requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
easyocr==1.7.1
filelock==3.15.4
fsspec==2024.6.1
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
imageio==2.35.1
Jinja2==3.1.4
jiter==0.5.0
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
openai==1.44.1
opencv-python-headless==4.10.0.84
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pydantic==2.9.1
pydantic_core==2.23.3
pytesseract==0.3.13
python-bidi==0.6.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
sniffio==1.3.1
sympy==1.13.2
tifffile==2024.8.30
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2